Moj pierwszy commit. Pozmienialem id na long +dodalem propertisa i porobilem foldery do jsp

+++ src/main/java/com/sda/javagda22/Pjeski/domain/repository/ShelterRepository.java
@@ -4,5 +4,5 @@ package com.sda.javagda22.Pjeski.domain.repository;
 import com.sda.javagda22.Pjeski.domain.model.Shelter;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ShelterRepository extends JpaRepository<Shelter, Integer> {
+public interface ShelterRepository extends JpaRepository<Shelter, Long> {
 }
